Add a couple of use statements to maintain access to string secondary methods
With the change on master to return to respecting the privacy and limitations
of use and import statements, these modules started encountering errors when
trying to access string.format (defined in IO) and string.splitMsgToTuple
(defined in ServerConfig).  For the former, add a use of IO to the main body
of the module as it is used in multiple functions.  For the latter, add a use
within the body of castMsg, as it is the only function in the module that uses
the method.

This allowed `make` to run to completion locally

----
Signed-off-by: Lydia Duncan <lydia-duncan@users.noreply.github.com>