Avoid converting GenSymEntry's to strings
There were some places that were converting a `GenSymEntry` into a
string. This resulted in converting the array contained in the symbol
entry into a string, which will be be very slow.

For now just print out the symbol entry name, which might not have been
the original intention, but resolves some large performance regressions
for now.

I'm not sure I caught all places that are doing this, but I think I
caught a majority.

Resolves 614