Make fill* random routines stable with varying numLocales
Previously `fillInt`, `fillUint`, `fillReal`, and `fillBool` would do a
coforall across the locales and use `fillRandom` on the  local array
with a seed that was based on `here.id`. Basing the seed on the locale
ID made the results unstable as the number of locales varies.

Instead of doing that, just use `fillRandom` on the distributed array.
This routine does similar operations under the covers on distributed
arrays as the manual coforall+local-fill, but we take care to make sure
the results are stable no matter how many locales are used.

This fixes the instability in `test_random_strings_uniform_with_seed()`
as the number of locales changes. I checked that test for 1-32 locales.