Optimize prod reductions
We want prod-reduce to return 0 if any element of the array is 0. This requires
special casing and checking for 0's because otherwise you can get into a
situation where you get a running `inf` and then `inf*0` will be `nan` instead
of 0.

Previously, for int64 this was achieved by doing an && reduction first to check
for 0's but this adds an extra reduction. Instead, use a forall with reduce
intents to do the prod-reduce and check for 0's with 1 pass over the array.

Use this for the float64 case as well, which was previously not checking for
0's.