Resolve a SipHash regression from 605
In 605 a param flag around the `TRACE` function was removed, which led to a
severe performance degradation for SipHash. What was previously a no-op,
became a `logger.debug()` call. And even though that logger.debug() call was
itself a no-op, there was still a lot of string creation and concatenation at
the call site. Add the param flag back to resolve the regression.

Performance on the nightly chapcs machine with comm=none:

Before:

```
make test-bin/SipHashSpeedTest
./test-bin/SipHashSpeedTest --NINPUTS=100_000
> Hashed 100000 blocks (6.0 MB) in 63.91 seconds (0.09 MB/s)
```

Now:

```
make test-bin/SipHashSpeedTest
./test-bin/SipHashSpeedTest --NINPUTS=100_000
> Hashed 100000 blocks (6.0 MB) in 0.00 seconds (17391.30 MB/s)
./test-bin/SipHashSpeedTest --NINPUTS=100_000_000
> Hashed 100000000 blocks (6152.0 MB) in 0.31 seconds (20112.27 MB/s)
```

Resolves 636