Use aggregation to gather values in string sorting
Use aggregation to gather values in string sorting. Elements from the
`values` array can be remote, so using aggregation avoids lots of
fine-grained remote accesses. To do that replace `unorderedCopy` for the
first `copyDigit` with assignment for local operations and aggregation
for the potentially remote `values`. Increase aggregation usage for the
second `copyDigit` by aggregating the remote `values` as well. Note that
this requires calculating and storing the bucket position prior to
gathering the new values/digits.