Convert unordredCopy to aggregation in findSegmentsMsg
`findSegmentsMsg` is the second half of `ak.GroupBy` and currently takes on the same order of time as the (co)argsort. The whole function could probably use a refactor, but one clear issue was a lingering `unorderedCopy` near the end, which this PR replaces with aggregation. This should hopefully benefit `GroupBy` on all platforms, but especially IB.