Update these operator declarations to use the operator keyword
Chapel 1.24 added support for declaring operator overloads with a new "operator"
keyword.  This enables operator definitions to be more easily found.
Declarations without the keyword will be deprecated starting in 1.25 (which is a
few months down the road, but I figured it would be good to address avoid the
deprecation warning as soon as possible)

Two of these could be operator methods.  I'm choosing to avoid doing that right
now for compatibility with 1.24 - there were some bugs with operator methods
remaining by the time of that release that should be resolved on master, or
shortly after a PR I have opened is merged.  In the future, we can choose to
use that functionality

----
Signed-off-by: Lydia Duncan <lydia-duncan@users.noreply.github.com>