Increase the aggregation buffer sizes for non-ugni configurations
Increase the aggregation buffer sizes from 4K to 8K for non-ugni
configurations. This is motivated by improving perforrmace on InfiniBand
systems, but the larger buffer sizes also benefit Ethernet networks as
well. Note that longer term we want to auto-tune these buffer sizes, but
for now this is a better default. On ugni increasing the size does not
change performance much so I didn't want to increase the memory overhead
for no benefit.

Here are some results for the nightly 16-node-cs-hdr configuration using
the default problem size:

| benchmark | before       | after        |
| --------: | -----------: | -----------: |
| gather    |  83.47 GiB/s | 108.17 GiB/s |
| scatter   | 106.99 GiB/s | 128.14 GiB/s |
| argsort   |   8.65 GiB/s |  10.08 GiB/s |