Fix categorical concatenate bug
Attempting to concatenate categorical arrays was resulting in scrambled category-code mappings. The root of this bug was that categories were being concatenated in unordered mode, whereas the follow-on logic assumed ordered categories. There should be a way to use ordered=False for the categories, but it would take a rewrite of the logic.