Optimize SrcAggregator copy
Optimize `SrcAggregator.copy()`. This switches from using chapel
arrays to c_ptr's for things that are locally aggregated in order to
clean up the generated code for the copy routine. This squashes some
wide pointers, which has the effect of eliminating some branches at
runtime. These branches are always correctly predicted so there's really
not a huge performance difference but I wanted to eliminate this is as a
possible source of overhead.

A more impactful change is to put an assert in a boundsChecking branch.
This assert was computing `here.id` each time, which is slightly
expensive and had some overhead with how often it's called.