Recommend Chapel 1.24.1 and use it for CI testing
Chapel 1.24.x contains a number of performance optimizations,
particularly on InfiniBand systems. For more information see
https://chapel-lang.org/releaseNotes/1.24/03a-perf-ib.pdf

Change the documentation to require 1.24.1, but only recommend it in the
Makefile instead of making it an error. Arkouda should continue to work
with Chapel 1.23 so existing users don't necessarily have to upgrade,
but I would recommend it for the performance gains.