Make the array values in groupby and aggregate benchmarks consistent
Previously, the values in the array were based on the array size, which
meant that larger values were used as the array size grew. Some
algorithms in Arkouda like radixsort are based on the number of bits
used so when we transitioned from something like 32-bit to 33-bit values
an extra round of radixsort would be required.

This gave us the false impression that there was a regression at certain
problem sizes when in reality it was about the array values not the
array size. To avoid this confusion just use 32-bit values for all
problem sizes. This matches what we do in all the benchmarks.

Resolves 793