Improve the performance of client-server array transfers
Improve the performance of `ak.array()` and `pdarray.to_ndarray()`. This
improves how the arrays are translated to bytes. For pdarray.to_ndarray,
previously we would just read from the distributed pdarray directly, but
this results in lots of fine-grained reads from that distributed array.
Change to copying to a local array, which takes advantage of Chapel's
bulk transfer, and then doing the byte conversion locally. For ak.array
do something similar but write into a local array and then do bulk
transfer to the distributed array.

This also adds a benchmark to track the performance. Note that unlike
most tests that take a per-locale size, this just takes a fixed size
since it ultimately depends on the client to head-node connection.

Here's the performance improvement for 16-node-xc:

| config | to_ndarray | ak.array   |
| ------ | ---------: | ---------: |
| before |  3.9 MiB/s |  4.7 MiB/s |
| after  | 20.5 MiB/s | 23.5 MiB/s |

And for 16-node-cs-hdr:

| config | to_ndarray | ak.array   |
| ------ | ---------: | ---------: |
| before |  2.4 MiB/s |  2.9 MiB/s |
| after  | 29.5 MiB/s | 25.9 MiB/s |

So roughly a ~5x improvement for XC and ~10x for CS.