Optimize to_ndarray by using np.frombuffer
Switch from using `struct.unpack` and creating a new array with copies
to just using `np.frombuffer`. Note that unfortuantely we still have to
make a copy because `bytes` aren't mutable so we have to create a
`bytearray` copy. There may be some way to eliminate this with
`memoryview`, but I'm not familiar with that yet.

For more info about `frombuffer` see
https://numpy.org/doc/stable/reference/generated/numpy.frombuffer.html

On an XC this improves to_ndarray performance from 20 MiB/s to 32 MiB/s
(39s to 24s)