Optimize how the pdarray is converted to bytes for to_ndarray
Instead of writing the localized array to a file and reading it back
into a bytes type, just directly create a new bytes type using one of
the factory functions and getting points to the underlying localized
array. This still results in a copy of the array's buffer to create the
bytes, but this is much faster than before (and we can eliminate that
copy by using the makeArrayFromPtr interface if we really want.)

Note that this breaks portability when the endianess of the server
doesn't match the client. Previously everything forced big-endian and
now we assume client and server match. This can probably be addressed by
sending the endianness along with the message and interpretting
differently, but I'm not positive.

On an XC this improves to_ndarray performance from 32 MiB/s to 360 MiB/s
(24s -> 2s)