Memoize the server config on the client
Memoize the server's configuration on calls to `ak.connect()` so that
`ak.get_config()` can return the server configuration without any server
additional calls. The motivation for this is to be able to check the
server byteorder/endianness in an upcoming client-server array transfer
optimization without an additional server call.