Add 04/06/21 scalability results for Cray XC and HPE Apollo
This includes benchmark runs on up to 240 nodes of a Cray XC with Aries
and an HPE Apollo with HDR-100 InfiniBand. I kept forgetting to commit
this data so it's from a few months ago, but the results are still
interesting.

For HPE Apollo we're comparing Chapel 1.23.0 with Arkouda from 10/20/20
(224a7c3) and Chapel 1.24.1 with Arkouda from 04/06/21 (795e7e2). There
were a number of InfiniBand optimizations between those Chapel and
Arkouda versions that led to significant performance and scalability
improvements.

For Cray XC we're just running with Chapel 1.24.1 and Arkouda from
04/06/21 and comparing that to the same configuration on Apollo. This
shows that modern Apollo systems can significantly outperform the aging
XC line. It's not really an apples-to-apples comparison because the core
counts are drastically different (36 vs 128) but you can't get this high
of a core count on an XC anyways, so I think it's a valid comparison.

There is data for stream, argsort, gather, and scatter using 8 GiB per
node and some larger argsort runs up to 256 GiB per node on Apollo.
Argsort has a fixed startup overhead based on the total number of cores.
This means the startup overhead is higher on the 128 core Apollo, but as
you increase the problem size the startup time matters less and the
overall rate improves significantly.