Avoid using pyzmq 22.3.0
Since pyzmq 22.3.0 was released we've been seeing testing failures for
registration_test.py due to unknown symbol errors (this is typically due
to a double free or early free of the symbol table entry.)

Until we have time to dig into this, pin to older pyzmq versions to see
if this resolves the nightly testing failures.

See 924 for details

While here, go back to using `start_test` for standalone Chapel CI tests
so that the error messages are visible when something goes wrong. The
config that runs the python tests is the slowest one anyways, so I don't
think having Chapel tests run slower will increase total CI time.