Optimize how segmented array slices are interpreted as strings/bytes
Optimize how slices of segmented arrays are interpreted as a Chapel
string/bytes. Interpreting is required to use Chapel's regex support,
which only operates on string/bytes types. Previously, the regex code
was localizing a region with a slice-assignment, and then making a new
string from it. This resulted in a large amount of short-lived
concurrent allocations, which was slow.

Update the regex code to use the new `lowLevelLocalizingSlice` routine
and instead of creating a new string, borrow memory from the segmented
array (or take ownership of the localized slice if the original data was
remote). This avoids allocation when the slice is local and minimizes
the amount of allocation and communication when the slice is remote.

This should significantly improve the performance of regex operations.

Part of 917
Part of 930