Fix sorting arrays with values on bitsPerDigit boundaries
Fix sorting arrays that contain a positive value that requires a
multiple of bitsPerDigit digits and at least one negative value. As an
optimization, the sort code checks how many bits are used so that we can
reduce the number of rounds of radix sort. e.g. if an array only
contains unsigned 16-bit values (`0 - 2**16-1`) we only need 1 round of
radixsort. Similarly, if an array only contains signed 16-bit values
(`-2**15 - 2**15-1`) we still only need 1 round of siphash.

However, there was a bug in the code where we were independently
computing the number of bits needed for positive and negative values.
If you have an array `0 - 2**16-1` you only need 16 bits, but as soon as
we add a negative value the signbit needs to be accounted for, which we
weren't doing before. So for an array with `-2**15 - 2**16-1` values we
were saying only 16 bits were needed even though it actually needs 17.
Fix that here by computing the number of bits needed for positive and
negative values and then if there are any negative values add 1 for the
signbit.

Resolves 937