Recommend Chapel 1.25 and use it for CI testing
Chapel 1.25 contains a number of performance optimizations and memory
fragmentation improvements, especially on InfiniBand systems. For more
information see:
 - https://chapel-lang.org/releaseNotes/1.25/05-perf.pdf
 - https://chapel-lang.org/releaseNotes/1.25/04-runtime.pdf

This changes the documentation to require 1.25, but only recommends it
in the Makefile instead of making it an error. Arkouda should continue
to work with Chapel 1.24 so existing users don't necessarily have to
upgrade, but I would recommend it for the performance gains.