Reduce argsort memory footprint
Reduce the memory overhead of argsort/radixSortLSD_ranks from ~5x to
~4x. The big arrays created are 2 temp arrays with the key and the rank
and a return array with just the rank. Previously all 3 arrays existed
at the same time for a ~5x overhead, but this refactors the code to
destroy one of the temp arrays before creating the return array, which
changes the peak overhead to those 2 temp arrays or a ~4x overhead.

While here remove some internal `overMemLimit` checks on a scan since
it's expected that callsites use radixSortLSD_memEst before calling
argsort and that includes the scan arrays.