Issue #914: Uses aggregator for SegString to HDF5 writes (#1016)
* Issue #914: Uses aggregator for SegString to HDF5 writes and
deprecates previous shuffling implementation.

* Issue #914: fixupSegBoundaries doesn't work when nOffsets < nLocales during HDF5write
and subsequent read.  This sets the default read/load operations on the client to use
calc_string_offsets=True

* Issue #914: Reverting change in client for calc_string_offsets in
favor of detection of nOffsets < nLocales in GenSymIO.
When there are empty domains for offsets in HDF5 files, we'll
revert to calculating offsets from values array, otherwise we'll
honor the segments and do the fixupSegBoundaries proc if specified by
client.