Speed up gasnet unit tests by only using 1 thread per locale (#1035)
Similar to 742, but for the chapel unit tests instead of the arkouda
pytests. At the time that PR was done the chapel unit tests were fast so
I didn't bother making this change, but they are currently the slowest
part of CI testing. This brings the gasnet unit tests down from 38 to 28
minutes and overall CI testing takes ~30 minutes.

Part of 388