Optimize small and medium int in1d operations (#1044)
Previously, int `in1d` had three implementation strategies depending on
the size of the second array. For small arrays (`<= 2**4`) a brute force
reduction was done, for medium arrays (`<= 2**25`) the array was
localized into a set, and for anything larger a more complicated
sorting-based scheme was used. Sorting has a lot of overhead so the
expectation was that the other strategies would be faster for smaller
problem sizes, but this wasn't tuned extensively before.

After adding a benchmark in 969, we discovered that the medium sized set
strategy was very slow in multi-locale configs. This was because the
code was serially gathering each element of the distributed array to add
to each locales private set. We know this fine-grained communication
performs poorly, especially on non Aries networks. Optimize that here by
doing bulk communication with the `lowLevelLocalizingSlice` to gather
the entire localSubdomain from each locale. This results in a single
bulk GET from each remote locale. We do this with an offset iterator so
that you don't have all locales doing a GETs from remote locales in
lockstep, which can overwhelm the network injection rate of a single
node. This was the most efficient strategy I found, and the results for
other schemes I found are in 970.

This significantly improves modest size in1d operations and brought
performance of really small operations close to the brute force scheme.
To close that gap I removed the PrivateDist array used for timers which
had small startup cost and with that the set based strategy is always
faster than the brute force strategy so I just removed the old brute
force method. Almost all (>95%) of the time in the set strategy is
serially adding to the set for non-trivial problem sizes so there's not
much user for the timers anyways now.

Here's some performance figures for various second array sizes on
16-node-cs-hdr before and after (first array size of `10**8`)

| Size      | Before      | Now         |
| --------- | ----------: | ----------: |
| 2**1      | 316.2 GiB/s | 324.7 GiB/s |
| 2**4      |  90.8 GiB/s | 250.1 GiB/s |
| 2**8      | 208.4 GiB/s | 247.3 GiB/s |
| 2**16     |  47.3 GiB/s | 177.3 GiB/s |
| 2**22     |   0.9 GiB/s |   9.4 GiB/s |
| 2**23     |   0.4 GiB/s |   4.8 GiB/s |
| 2**24     |   0.2 GiB/s |   2.5 GiB/s |
| 2**25 - 1 |   0.1 GiB/s |   1.3 GiB/s |
| 2**25 + 1 |   4.2 GiB/s |   4.2 GiB/s |

So overall some large performance benefits, though we'll probably still
want to do some tuning on the threshold between the medium and large
strategy.

Given that the medium size is doing an all gather, I wanted to check
scaling too. Here's some 512-node-xc results:

| Size      | Before      | Now         |
| --------- | ----------: | ----------: |
| 2**25 - 1 |   0.4 GiB/s |   6.4 GiB/s |
| 2**25 + 1 |   4.1 GiB/s |   4.1 GiB/s |

So here we still see improved performance for the set strategy, and in
this case it's actually faster than the sort strategy at the current
threshold.

The main bottleneck in the set strategy is that we're serially adding to
a set and this serialization becomes more apparent on higher core count
machines. It's unlikely that we'll be able to come up with a perfect
simple threshold because it depends on core count and network speed, but
I would still lean towards dropping it to `2**22` or `2**23` or so.

This threshold is also optimization dependent and there's likely more we
can do to optimize both strategies. Chapel should eventually have a
portable and scalable concurrent set which would help with the current
serialization. I haven't looked at tuning the sort-based strategy at all
yet but there is likely room for improvement. In the end, I don't think
we need a prefect threshold and I lean towards lowering the sort
threshold because sort is used for so much else it has well known
performance behavior.

Part of 970