Rework IO benchmarks to ues the same benchmarks for Parquet and HDF5 Today, the HDF5 and Parquet benchmarks are nearly identical, just with the different Parquet/HDF5 API calls. This PR combines them both into a single benchmark that defaults to HDF5 and then adds a separate Parquet version with just a main function and parser.
This also allows the IO benchmarks to run just write/read tests
if you only wish to test one of the 2 and wanted to keep the files
around afterwards, for example.

Additionally, a benchmark is added to test writing/reading multiple
files, which represents a more common use case and illustrates how
Parquet and HDF5 stack up for multiple file reading, which seems to
be a more realistic use case that is also made more interesting
after having parallelized Parquet file reads.