Reduce argsort memory overhead (#1061)
* Reduce argsort memory overhead

Reduce the memory overhead of `argsort`/`radixSortLSD_ranks` from ~5x to
~4x. Previously, `radixSortLSD_ranks` had 3 large arrays in existence at
the same time -- 2 temp key+rank arrays and a return rank array. This
creates a `radixSortLSD` wrapper that returns the key+ranks, and updates
`radixSortLSD_ranks` to call that. This means we only have 1 temp array
when we create the ranks arrays, which reduces the peak memory to the 2
temp key+ranks arrays.

I created `radixSortLSD` instead of just adjusting `radixSortLSD_ranks`
to destroy one of the temp arrays early since we can use it to optimize
in1d, which needs the sorted keys and the ranks, but today does that
inefficiently with a rank sort and gathering values into a sorted array.

* Update radix sort memory usage estimate to reflect reduced usage

Update `radixSortLSD_memEst` to reflect the reduced memory usage and use
it in additional places that were duplicating the calculation before.