Optimize sorting calls when both keys and ranks are needed (#1063)
Previously, code that needed both the sorted keys and the ranks was
doing a rank sort and then gathering to sort the keys. Sorting the keys
is already done as part of the main radixSort code, so call the new
version that returns both the keys and values and then split that into 2
arrays. In some cases it would be pretty easy to just use the combined
array, but I wasn't sure if that was worthwhile.

This also removes some explicit `isSorted` checks since the sort code
will already do that check and removes an unused `assumeSorted` from
`uniqueGroup`

On 16-node-cs-hdr I see large int in1d go from 4.8 to 5.4 GiB/s and
setdiff go from 3.6 to 4.1 GiB/s. I expect some improvements for codes
using the other areas that were optimized.

Resolves 1062
Part of 970