Add a --saveUsedModules flag to save modules used in a server run (#1065)
This PR adds a --saveUsedModules flag that prints the name of all
modules that had a mesasge function called in a UsedModules.cfg file
that can be reused to speed up the next build. The function is called
upon `ak.shutdown()`.

To use this flag on a benchmark, it can be run:
`./benchmarks/run-benchmarks.py IO --server-args='--saveUsedModules`.
To use this flag on an interactive run of the Arkouda server, run:
`./arkouda_server --saveUsedModules`.