Switch `SymEntry` to use const domains
Arkouda arrays are never resized so there's no reason for the domains to
not be const. Making them const allows Chapel to skip some tracking code
that's used for potential array resizing, so this slightly optimizes
array creation and deletion.

For more information on this optimization see the content starting on
slide 25 of https://chapel-lang.org/releaseNotes/1.23/04-perf-opt.pdf.

Resolves 1083