Fix `computeOnSegments` when there's more locales than strings
Many tests were failing in nightly 16 locale testing after 1060. This
was caused by trying to hash fewer strings than there were locales. The
empty locales should be a no-op, but ended up trying to do a count with
a negative value, which isn't supported on ranges without a high bound.
Fix that by just making the number of segments 0 if it was negative.

Alternatively we could skip further computation if the size is 0, but
that adds an extra level of indentation, which didn't seem worthwhile.
The operations will be no-ops anyways so it shouldn't matter.

Resolves 1093