Re-enable python portability testing
This was originally disabled because we were seeing a lot of unknown
symbol errors for it. These errors were caused by garbage collection
collecting references to stale arrays that shared the same name on a new
server. We likely saw it more frequently in the portability testing mode
because GC behavior can vary across python versions. Now that the core
issue has been resolved we can restart the portability testing.

See 924 for more info
Resolves 694