Have each arkouda server use a unique identifier for arrays (#1096)
* Have each arkouda server use a unique identifier for arrays

Previously, a server would just generate names like `id_1` and use a
monotonically increasing counter so that each server instance had unique
ids. This works well for a single client/server session, but can cause
problems when a single python instance connects to multiple servers. In
this case it's possible that stale references to an old server are only
garbage-collected when connected to a new server. This can cause
accidental deletion of live arrays on the new server. This has
occasionally been an issue for users and it has been sporadically biting
us in CI testing, which took us a while to figure out.

Switch to generating a short random token to add to the ids so each
server should have a unique identifier. e.g. `id_a4e9f43_1`. This way
connections to a new server can't delete ids acquired from an old
server.

Resolves 924

* Revert "Avoid using pyzmq 22.3.0"

This reverts commit 0ba01be1b78609976ec8e2c3d6a0d492215bc3fe.

We stopped using pyzmq 22.3.0 since it appeared to introduce some
unknown symbol errors in CI testing, but that was a red-herring. This
issue has been correctly addressed by giving each server a unique id, so
we no longer need to pin to older versions.

See 924 for more info

* Stop manually registering some arrays in regex_test.py

This was a workaround for the sporadic unknown symbol error, but we have
since fixed the core issue.

See 924 for more info.

* Re-enable python portability testing

This was originally disabled because we were seeing a lot of unknown
symbol errors for it. These errors were caused by garbage collection
collecting references to stale arrays that shared the same name on a new
server. We likely saw it more frequently in the portability testing mode
because GC behavior can vary across python versions. Now that the core
issue has been resolved we can restart the portability testing.

See 924 for more info
Resolves 694